🎨 重做的毕业度计算

+++ GenshinUID/genshinuid_enka/mono/Power.py
@@ -7,7 +7,7 @@ class Power(BaseModel):
     raw: str
     percent: float
     value: float
-    plus: int
+    plus: float
 
 
 class sp_prop(BaseModel):
